ai-portfolio-builder/platform/backend/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.3
yfinance>=0.2
feedparser>=6.0
anthropic>=0.89
langgraph>=0.2
langchain-core>=0.3
langchain-anthropic>=0.3
